:bug: occurs afterr creating a report if excel is not installed

+++ src/peakfinding/groupby/emutils.h
@@ -35,7 +35,7 @@ namespace peakfinding{
 	double pdfparam(blas::vector<double> ,blas::vector<double>);
 	double logpdfparam(blas::vector<double> ,blas::vector<double>);
 	void   oneemstep(matrix&, matrix&, matrix&,double);
-	void   emsteps(matrix&, matrix&, matrix&,unsigned,double,double);
+	void   emsteps(matrix&, matrix&, matrix&,size_t,double,double);
 	double scoreparam(blas::vector<double> ,blas::vector<double>);
 	matrix scoreparams(const matrix &, const matrix &);
 	matrix logscoreparams(const matrix &, const matrix &);
